perf(fs): default classes_dtype int32->int16 + thread it through, halving the (n,K) FE-MI code buffer
The batched FE-MI + noise-gate kernel densifies each candidate column into a (n, K) ``classes_dense``
buffer. Its dense codes are in [0, n_dense) <= n_bins, so int16 holds any realistic nbins (<32768) and
HALVES this buffer vs the prior int32 default -- value-identical, because the codes are read-only
histogram indices (joint_counts[classes_dense[r,k], ...]) whose storage width never changes a count.

- batch_mi_with_noise_gate: default classes_dtype int32 -> int16.
- _pairs_dispatch (3 call sites): cap the densified width at int16 (disc_2d.dtype if itemsize<=2 else
  int16) so wide disc inputs no longer fall back to int32; narrow int8 disc still passes int8 and wins more.
- _gpu_resident_fe grand-fused CPU fallback: classes_dtype int16.

Bit-identity verified: int8 == int16 == int32 == int64 fe_mi (array_equal) on the canonical batch.
gpu_resident suite 13/13. Production already passed int8 (from the int8 disc) so its path is unchanged;
this narrows every caller that relied on the old int32 default (sweeps, tests, wide-disc paths).

+++ src/mlframe/feature_selection/filters/_gpu_resident_fe.py
@@ -458,7 +458,7 @@ def grand_fused_pair_mi(
                 disc_2d=disc_host, factors_nbins=fnb, classes_y=y_i64, classes_y_safe=csafe, freqs_y=fy,
                 npermutations=int(npermutations), base_seed=np.uint64(0),
                 min_nonzero_confidence=float(min_nonzero_confidence), use_su=bool(use_su),
-                dtype=np.int32, classes_dtype=np.int32,
+                dtype=np.int32, classes_dtype=np.int16,
             )
         else:
             fe_mi = out[0] if isinstance(out, tuple) else out
